+++ packages/core/resolve-core/.gitignore
@@ -1,0 +1,1 @@
+!src/types
