+++ packages/core/resolve-core/src/aggregate/validate-aggregate.ts
@@ -1,4 +1,4 @@
-import { AggregateMeta } from '../type-definitions/runtime'
+import { AggregateMeta } from '../types/runtime'
 
 // FIXME: add validation (io-ts?)
 export const validateAggregate = (aggregate: AggregateMeta) => aggregate
