+++ packages/core/resolve-core/src/index.ts
@@ -1,6 +1,6 @@
 export * from './helpers'
-export * from './type-definitions/core'
-export * from './type-definitions/runtime'
+export * from './types/core'
+export * from './types/runtime'
 export * from './symbols'
 export * from './errors'
 export * from './interop'
