+++ packages/core/resolve-core/src/read-model/validate-read-model.ts
@@ -1,4 +1,4 @@
-import { ReadModelMeta } from '../type-definitions/runtime'
+import { ReadModelMeta } from '../types/runtime'
 
 // FIXME: add validation (io-ts?)
 export const validateReadModel = (readModel: ReadModelMeta) => readModel
