+++ packages/core/resolve-core/src/saga/validate-saga.ts
@@ -1,4 +1,4 @@
-import { SagaMeta } from '../type-definitions/runtime'
+import { SagaMeta } from '../types/runtime'
 
 // FIXME: add validation (io-ts?)
 export const validateSaga = (saga: SagaMeta) => saga
