+++ packages/core/resolve-core/src/types/core.ts
